Rewrote the inits and resets of the ExecutionAlgo, TWAPAlgo and Broker class
Previously, everytime the algo was simulated the broker was reset, which resets the Algo as well. This lead to different buckets and hence completely different backtest results. Now the reset simply sets simulation counters to 0 and simulating the same algo always yield the same result, the buckets are calculated during the __init__