Fixing the state sampling after calculating rewards.
Right now, no matter how many times you evaluate the benchmark_algo, the results are consistent. However, if you simulate the rl_algo with the same parameters as the benchmark_algo, the size of the trade logs and the vwap are different, which doesn't make sense, different LOBs are being sampled, need to figure out why.