Rounding errors due to using Decimal values/ error with negative durations, start_time and end_time
The commented out lines raise errors due to rounding, in the scale of the tick_size. I suggest commenting this comprobations out since these errors are inconsequential.

A second bug has been fixed in which the start time was at the end of a day, but the end time wasn't properly calculated as belonging to the following day (this will potentially be important once we add more data).