Update base_env_real.py
Quick fix to the reset() function, before the obs_space wasn't reset with the BaseEnv and this could lead to higher volumes/orders left in the buckets in the Obs than those allowed in the obs_space. Now the limits are set to inf to not have to reset the obs_space at every episode end.