Added: Outperformance vs the Benchmark Algo Reward Metric calculated at the end of each bucket, Vectorized the VWAP calculation.
Now the Algo simulations can be carried out at the end of each bucket, and rewards based on the perofrmance on each bucket can be then calculated. The performance calculation of the vwap has also been vectorized.

Issues: The rendering happens at the beginning of the 10th episode, when the volumes_per_trade hasn't been populated yet, and hence the rl_algo submits only 1 market order that often doesn't fully execute and the results look weird as a result. Fix: How to make it so that the rendering happens at the end of the episode? Also I haven't managed to run the test_agent_one_episode because the RNN is lacking an input (was this working for you Florin?)