Bugfix in relation to state construction
The attribute algo_events carries all the events of the algo, thus at a bucket end, we were selecting the timestamp of the bucket_end rather than that of the next order_placement. This didn't lead to any issue when training because states were produced regardless, but it would lag the information the algo should be getting.